+++ packages/zowe-explorer/__tests__/__unit__/dataset/actions.unit.test.ts
@@ -21,6 +21,7 @@ import {
     createTextDocument,
     createTreeView,
     createQuickPickContent,
+    createWorkspaceConfiguration,
 } from "../../../__mocks__/mockCreators/shared";
 import {
     createDatasetAttributes,
@@ -2903,7 +2904,11 @@ describe("Dataset Actions Unit Tests - Function createFile", () => {
         await dsActions.createFile(blockMocks.datasetSessionNode, blockMocks.testDatasetTree);
         expect(mocked(vscode.workspace.getConfiguration)).lastCalledWith(globals.SETTINGS_DS_DEFAULT_PS);
 
-        expect(createDataSetSpy).toHaveBeenCalledTimes(5);
+        mocked(vscode.window.showQuickPick).mockResolvedValueOnce("Partitioned Data Set: Extended" as any);
+        await dsActions.createFile(blockMocks.datasetSessionNode, blockMocks.testDatasetTree);
+        expect(mocked(vscode.workspace.getConfiguration)).lastCalledWith(globals.SETTINGS_DS_DEFAULT_EXTENDED);
+
+        expect(createDataSetSpy).toHaveBeenCalledTimes(6);
     });
     it("Checking of proper configuration being picked up for different DS types with credentials prompt", async () => {
         createGlobalMocks();
@@ -2950,7 +2955,11 @@ describe("Dataset Actions Unit Tests - Function createFile", () => {
         await dsActions.createFile(node, blockMocks.testDatasetTree);
         expect(mocked(vscode.workspace.getConfiguration)).lastCalledWith(globals.SETTINGS_DS_DEFAULT_PS);
 
-        expect(createDataSetSpy).toHaveBeenCalledTimes(5);
+        mocked(vscode.window.showQuickPick).mockResolvedValueOnce("Partitioned Data Set: Extended" as any);
+        await dsActions.createFile(node, blockMocks.testDatasetTree);
+        expect(mocked(vscode.workspace.getConfiguration)).lastCalledWith(globals.SETTINGS_DS_DEFAULT_EXTENDED);
+
+        expect(createDataSetSpy).toHaveBeenCalledTimes(6);
     });
     it("Checking of proper configuration being picked up for different DS types with credentials prompt for favorite", async () => {
         createGlobalMocks();
@@ -2996,7 +3005,11 @@ describe("Dataset Actions Unit Tests - Function createFile", () => {
         await dsActions.createFile(node, blockMocks.testDatasetTree);
         expect(mocked(vscode.workspace.getConfiguration)).lastCalledWith(globals.SETTINGS_DS_DEFAULT_PS);
 
-        expect(createDataSetSpy).toHaveBeenCalledTimes(5);
+        mocked(vscode.window.showQuickPick).mockResolvedValueOnce("Partitioned Data Set: Extended" as any);
+        await dsActions.createFile(node, blockMocks.testDatasetTree);
+        expect(mocked(vscode.workspace.getConfiguration)).lastCalledWith(globals.SETTINGS_DS_DEFAULT_EXTENDED);
+
+        expect(createDataSetSpy).toHaveBeenCalledTimes(6);
         createDataSetSpy.mockClear();
     });
     it("Checking PS dataset creation", async () => {
@@ -3604,5 +3617,36 @@ describe("Dataset Actions Unit Tests - Function confirmJobSubmission", () => {
             title: "Submit",
         });
         await expect(dsActions.confirmJobSubmission("Profile\\test.jcl", true)).resolves.toEqual(true);
+    });
+});
+
+describe("Dataset Actions Unit Tests - Function getDsTypePropertiesFromWorkspaceConfig", () => {
+    it("Should use use local JCL doc name for confirmJobSubmission", async () => {
+        createGlobalMocks();
+        const options = createWorkspaceConfiguration();
+        // const opt = dsActions.getDsTypePropertiesFromWorkspaceConfig(options);
+        jest.spyOn(options, "get").mockImplementation((attr: string) => {
+            const pdse = {
+                dsntype: "LIBRARY",
+                dsorg: "PO",
+                alcunit: "CYL",
+                primary: 10,
+                secondary: 3,
+                dirblk: 25,
+                recfm: "FB",
+                blksize: 27920,
+                lrecl: 80,
+            };
+            return pdse[attr];
+        });
+        expect(dsActions.getDsTypePropertiesFromWorkspaceConfig(options).dsntype).toBe("LIBRARY");
+        expect(dsActions.getDsTypePropertiesFromWorkspaceConfig(options).dsorg).toBe("PO");
+        expect(dsActions.getDsTypePropertiesFromWorkspaceConfig(options).alcunit).toBe("CYL");
+        expect(dsActions.getDsTypePropertiesFromWorkspaceConfig(options).primary).toBe(10);
+        expect(dsActions.getDsTypePropertiesFromWorkspaceConfig(options).secondary).toBe(3);
+        expect(dsActions.getDsTypePropertiesFromWorkspaceConfig(options).dirblk).toBe(25);
+        expect(dsActions.getDsTypePropertiesFromWorkspaceConfig(options).recfm).toBe("FB");
+        expect(dsActions.getDsTypePropertiesFromWorkspaceConfig(options).blksize).toBe(27920);
+        expect(dsActions.getDsTypePropertiesFromWorkspaceConfig(options).lrecl).toBe(80);
     });
 });